move Persistence tests
