+++ tests/Persistence/SqlTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Atk4\Data\Tests;
+namespace Atk4\Data\Tests\Persistence;
 
 use Atk4\Data\Exception;
 use Atk4\Data\Model;
 use Atk4\Data\Schema\TestCase;
 
-class PersistentSqlTest extends TestCase
+class SqlTest extends TestCase
 {
     public function testLoadArray(): void
     {
