+++ tests/Util/DeepCopyTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Atk4\Data\Tests;
+namespace Atk4\Data\Tests\Util;
 
 use Atk4\Data\Model;
 use Atk4\Data\Schema\TestCase;
